pyproject.toml
[tool.poetry]
name = "difference-viewer"
version = "0.1.0"
description = "A graphical tool for comparing documents"
authors = ["Blueno <k.pg.blueno@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.8.8"
pyqt5 = "^5.15.1"
pyqt5-tools = "^5.15.1.1.7.5"
pywin32 = "^300"
opencv-python = "^4.4.0.46"
numpy = "^1.20.1"
pillow = "^8.1.2"
python-pptx = "^1.0.2"
xdwlib = "^3.10.9"
pymupdf = "^1.24.11"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
